move common funcs to common::benchmarking
adjust to the change. Fix build of runtime
