+++ common/Cargo.toml
@@ -22,7 +22,10 @@ sp-runtime = { version = "5.0.0", git = "https://github.com/gear-tech/substrate.
 sp-std = { version = "4.0.0", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
 sp-arithmetic = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
 frame-support = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
+frame-system = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false, optional = true }
+frame-benchmarking = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false, optional = true }
 gear-runtime-interface = { path = "../runtime-interface", default-features = false }
+parity-wasm = { version = "0.42.2", default-features = false, optional = true }
 
 [features]
 default = ["std"]
@@ -37,4 +40,10 @@ std = [
     "frame-support/std",
     "primitive-types/std",
     "gear-runtime-interface/std",
+]
+runtime-benchmarks = [
+    "frame-benchmarking",
+    "frame-system/runtime-benchmarks",
+    "frame-support/runtime-benchmarks",
+    "parity-wasm",
 ]
